requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
antlr4-python3-runtime==4.13.0
antlr4-tools==0.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.3
flatdict==4.0.1
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
install-jdk==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
junit2html==30.1.3
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
networkx==3.1
numpy==1.24.2
ortools==9.6.2534
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
plotly==5.15.0
pluggy==1.0.0
prometheus-client==0.17.1
protobuf==4.22.1
py==1.11.0
pyparsing==3.0.9
pytest==7.2.2
pytest-asyncio==0.21.1
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
responses==0.23.1
scipy==1.10.1
six==1.16.0
tenacity==8.2.2
tqdm==4.65.0
types-PyYAML==6.0.12.10
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.2.3
yarl==1.9.2
